+++ conversation_service/models/conversation_models.py
@@ -1,8 +1,10 @@
+"""Pydantic models for conversation requests and responses."""
+
 from __future__ import annotations
 
-from pydantic import BaseModel, ConfigDict, Field, field_validator, model_validator
 from uuid import UUID
-from pydantic import BaseModel, Field, ConfigDict, field_validator
+
+from pydantic import BaseModel, ConfigDict, Field, field_validator
 
 
 class ConversationMetadata(BaseModel):
@@ -12,28 +14,8 @@ class ConversationMetadata(BaseModel):
     confidence_score: float | None = Field(default=None, ge=0.0, le=1.0)
 
     model_config = ConfigDict(json_schema_extra={
-        "example": {
-            "intent": "greeting",
-            "confidence_score": 0.95,
-        }
+        "example": {"intent": "greeting", "confidence_score": 0.95}
     })
-
-    @field_validator("user_id")
-    @classmethod
-    def user_id_positive(cls, v: int) -> int:
-        if v <= 0:
-            raise ValueError("user_id must be positive")
-        return v
-
-    @model_validator(mode="after")
-    def validate_consistency(self) -> "Conversation":
-        if self.total_turns > self.max_turns:
-            raise ValueError("total_turns cannot exceed max_turns")
-        if self.updated_at < self.created_at:
-            raise ValueError("updated_at must be after created_at")
-        if self.last_activity_at < self.created_at:
-            raise ValueError("last_activity_at must be after created_at")
-        return self
 
 
 class ConversationContext(BaseModel):
@@ -48,19 +30,6 @@ class ConversationContext(BaseModel):
             "turn_number": 1,
         }
     })
-
-    @field_validator("conversation_id", "start_turn", "end_turn")
-    @classmethod
-    def positive_numbers(cls, v: int) -> int:
-        if v <= 0:
-            raise ValueError("must be positive")
-        return v
-
-    @model_validator(mode="after")
-    def validate_turns(self) -> "ConversationSummary":
-        if self.end_turn < self.start_turn:
-            raise ValueError("end_turn must be >= start_turn")
-        return self
 
 
 class ConversationRequest(BaseModel):
@@ -112,25 +81,6 @@ class ConversationResponse(BaseModel):
         }
     })
 
-    @field_validator("conversation_id", "turn_number", "search_results_count")
-    @classmethod
-    def non_negative(cls, v: int) -> int:
-        if v < 0:
-            raise ValueError("must be non-negative")
-        return v
-
-    @field_validator("processing_time_ms", "search_execution_time_ms")
-    @classmethod
-    def positive_floats(cls, v: Optional[float]) -> Optional[float]:
-        if v is not None and v < 0:
-            raise ValueError("must be non-negative")
-        return v
-
-    @model_validator(mode="after")
-    def validate_dates(self) -> "ConversationTurn":
-        if self.updated_at < self.created_at:
-            raise ValueError("updated_at must be after created_at")
-        return self
     @field_validator("language")
     @classmethod
     def validate_language(cls, v: str) -> str:
